+++ doc/code/torch.rst
@@ -1,4 +1,4 @@
-.. automodule:: pennylane.torch
+.. automodule:: pennylane.interfaces.torch
    :members:
    :private-members:
    :special-members:
